requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
xlrd==1.2.0
pandas==1.1.4
django-rest-swagger
requests